advocate data retrieval moved from client to server + client functionality moved to dashboard component
