+++ .env
@@ -1,1 +1,2 @@
 DATABASE_URL=postgresql://postgres:password@localhost/solaceassignment
+NEXT_PUBLIC_LOCAL_URL=http://localhost:3000
