+++ src/app/dashboard.tsx
@@ -1,0 +1,87 @@
+"use client";
+
+import { ChangeEvent, useState } from "react";
+import type { Advocate } from "@/types";
+
+type Props = {
+  advocates: Advocate[];
+};
+
+export default function Dashboard({ advocates }: Props) {
+  const [filteredAdvocates, setFilteredAdvocates] = useState(advocates);
+
+  const onChange = (e: ChangeEvent<HTMLInputElement>) => {
+    const searchTerm = e.target.value;
+
+    document.getElementById("search-term")!.innerHTML = searchTerm;
+
+    console.log("filtering advocates...");
+    const filteredAdvocates = advocates.filter((advocate: Advocate) => {
+      return (
+        advocate.firstName.includes(searchTerm) ||
+        advocate.lastName.includes(searchTerm) ||
+        advocate.city.includes(searchTerm) ||
+        advocate.degree.includes(searchTerm) ||
+        advocate.specialties.includes(searchTerm) ||
+        String(advocate.yearsOfExperience) === searchTerm
+      );
+    });
+
+    setFilteredAdvocates(filteredAdvocates);
+  };
+
+  const onClick = () => {
+    console.log(advocates);
+    setFilteredAdvocates(advocates);
+  };
+
+  return (
+    <main style={{ margin: "24px" }}>
+      <h1>Solace Advocates</h1>
+      <br />
+      <br />
+      <div>
+        <p>Search</p>
+        <p>
+          Searching for: <span id="search-term"></span>
+        </p>
+        <input style={{ border: "1px solid black" }} onChange={onChange} />
+        <button onClick={onClick}>Reset Search</button>
+      </div>
+      <br />
+      <br />
+      <table>
+        <thead>
+          <tr>
+            <th>First Name</th>
+            <th>Last Name</th>
+            <th>City</th>
+            <th>Degree</th>
+            <th>Specialties</th>
+            <th>Years of Experience</th>
+            <th>Phone Number</th>
+          </tr>
+        </thead>
+        <tbody>
+          {filteredAdvocates.map((advocate: Advocate) => {
+            return (
+              <tr key={advocate.id}>
+                <td>{advocate.firstName}</td>
+                <td>{advocate.lastName}</td>
+                <td>{advocate.city}</td>
+                <td>{advocate.degree}</td>
+                <td>
+                  {advocate.specialties.map((s, i) => (
+                    <div key={i}>{s}</div>
+                  ))}
+                </td>
+                <td>{advocate.yearsOfExperience}</td>
+                <td>{advocate.phoneNumber}</td>
+              </tr>
+            );
+          })}
+        </tbody>
+      </table>
+    </main>
+  );
+}
